refactor(aiCore): 提取MixedBaseAPIClient基类重构API客户端
将AihubmixAPIClient和NewAPIClient的公共逻辑提取到MixedBaseAPIClient基类
减少代码重复，提高可维护性
