+++ src/renderer/src/aiCore/clients/AihubmixAPIClient.ts
@@ -1,43 +1,23 @@
 import { isOpenAILLMModel } from '@renderer/config/models'
-import {
-  GenerateImageParams,
-  MCPCallToolResponse,
-  MCPTool,
-  MCPToolResponse,
-  Model,
-  Provider,
-  ToolCallResponse
-} from '@renderer/types'
-import {
-  RequestOptions,
-  SdkInstance,
-  SdkMessageParam,
-  SdkModel,
-  SdkParams,
-  SdkRawChunk,
-  SdkRawOutput,
-  SdkTool,
-  SdkToolCall
-} from '@renderer/types/sdk'
+import { Model, Provider } from '@renderer/types'
 
-import { CompletionsContext } from '../middleware/types'
 import { AnthropicAPIClient } from './anthropic/AnthropicAPIClient'
 import { BaseApiClient } from './BaseApiClient'
 import { GeminiAPIClient } from './gemini/GeminiAPIClient'
+import { MixedBaseAPIClient } from './MixedBaseApiClient'
 import { OpenAIAPIClient } from './openai/OpenAIApiClient'
 import { OpenAIResponseAPIClient } from './openai/OpenAIResponseAPIClient'
-import { RequestTransformer, ResponseChunkTransformer } from './types'
 
 /**
  * AihubmixAPIClient - 根据模型类型自动选择合适的ApiClient
  * 使用装饰器模式实现，在ApiClient层面进行模型路由
  */
-export class AihubmixAPIClient extends BaseApiClient {
+export class AihubmixAPIClient extends MixedBaseAPIClient {
   // 使用联合类型而不是any，保持类型安全
-  private clients: Map<string, AnthropicAPIClient | GeminiAPIClient | OpenAIResponseAPIClient | OpenAIAPIClient> =
+  protected clients: Map<string, AnthropicAPIClient | GeminiAPIClient | OpenAIResponseAPIClient | OpenAIAPIClient> =
     new Map()
-  private defaultClient: OpenAIAPIClient
-  private currentClient: BaseApiClient
+  protected defaultClient: OpenAIAPIClient
+  protected currentClient: BaseApiClient
 
   constructor(provider: Provider) {
     super(provider)
@@ -74,23 +54,9 @@ export class AihubmixAPIClient extends BaseApiClient {
   }
 
   /**
-   * 类型守卫：确保client是BaseApiClient的实例
-   */
-  private isValidClient(client: unknown): client is BaseApiClient {
-    return (
-      client !== null &&
-      client !== undefined &&
-      typeof client === 'object' &&
-      'createCompletions' in client &&
-      'getRequestTransformer' in client &&
-      'getResponseChunkTransformer' in client
-    )
-  }
-
-  /**
    * 根据模型获取合适的client
    */
-  private getClient(model: Model): BaseApiClient {
+  protected getClient(model: Model): BaseApiClient {
     const id = model.id.toLowerCase()
 
     // claude开头
@@ -126,115 +92,5 @@ export class AihubmixAPIClient extends BaseApiClient {
     }
 
     return this.defaultClient as BaseApiClient
-  }
-
-  /**
-   * 根据模型选择合适的client并委托调用
-   */
-  public getClientForModel(model: Model): BaseApiClient {
-    this.currentClient = this.getClient(model)
-    return this.currentClient
-  }
-
-  /**
-   * 重写基类方法，返回内部实际使用的客户端类型
-   */
-  public override getClientCompatibilityType(model?: Model): string[] {
-    if (!model) {
-      return [this.constructor.name]
-    }
-
-    const actualClient = this.getClient(model)
-    return actualClient.getClientCompatibilityType(model)
-  }
-
-  // ============ BaseApiClient 抽象方法实现 ============
-
-  async createCompletions(payload: SdkParams, options?: RequestOptions): Promise<SdkRawOutput> {
-    // 尝试从payload中提取模型信息来选择client
-    const modelId = this.extractModelFromPayload(payload)
-    if (modelId) {
-      const modelObj = { id: modelId } as Model
-      const targetClient = this.getClient(modelObj)
-      return targetClient.createCompletions(payload, options)
-    }
-
-    // 如果无法从payload中提取模型，使用当前设置的client
-    return this.currentClient.createCompletions(payload, options)
-  }
-
-  /**
-   * 从SDK payload中提取模型ID
-   */
-  private extractModelFromPayload(payload: SdkParams): string | null {
-    // 不同的SDK可能有不同的字段名
-    if ('model' in payload && typeof payload.model === 'string') {
-      return payload.model
-    }
-    return null
-  }
-
-  async generateImage(params: GenerateImageParams): Promise<string[]> {
-    return this.currentClient.generateImage(params)
-  }
-
-  async getEmbeddingDimensions(model?: Model): Promise<number> {
-    const client = model ? this.getClient(model) : this.currentClient
-    return client.getEmbeddingDimensions(model)
-  }
-
-  async listModels(): Promise<SdkModel[]> {
-    // 可以聚合所有client的模型，或者使用默认client
-    return this.defaultClient.listModels()
-  }
-
-  async getSdkInstance(): Promise<SdkInstance> {
-    return this.currentClient.getSdkInstance()
-  }
-
-  getRequestTransformer(): RequestTransformer<SdkParams, SdkMessageParam> {
-    return this.currentClient.getRequestTransformer()
-  }
-
-  getResponseChunkTransformer(ctx: CompletionsContext): ResponseChunkTransformer<SdkRawChunk> {
-    return this.currentClient.getResponseChunkTransformer(ctx)
-  }
-
-  convertMcpToolsToSdkTools(mcpTools: MCPTool[]): SdkTool[] {
-    return this.currentClient.convertMcpToolsToSdkTools(mcpTools)
-  }
-
-  convertSdkToolCallToMcp(toolCall: SdkToolCall, mcpTools: MCPTool[]): MCPTool | undefined {
-    return this.currentClient.convertSdkToolCallToMcp(toolCall, mcpTools)
-  }
-
-  convertSdkToolCallToMcpToolResponse(toolCall: SdkToolCall, mcpTool: MCPTool): ToolCallResponse {
-    return this.currentClient.convertSdkToolCallToMcpToolResponse(toolCall, mcpTool)
-  }
-
-  buildSdkMessages(
-    currentReqMessages: SdkMessageParam[],
-    output: SdkRawOutput | string,
-    toolResults: SdkMessageParam[],
-    toolCalls?: SdkToolCall[]
-  ): SdkMessageParam[] {
-    return this.currentClient.buildSdkMessages(currentReqMessages, output, toolResults, toolCalls)
-  }
-
-  convertMcpToolResponseToSdkMessageParam(
-    mcpToolResponse: MCPToolResponse,
-    resp: MCPCallToolResponse,
-    model: Model
-  ): SdkMessageParam | undefined {
-    const client = this.getClient(model)
-    return client.convertMcpToolResponseToSdkMessageParam(mcpToolResponse, resp, model)
-  }
-
-  extractMessagesFromSdkPayload(sdkPayload: SdkParams): SdkMessageParam[] {
-    return this.currentClient.extractMessagesFromSdkPayload(sdkPayload)
-  }
-
-  estimateMessageTokens(message: SdkMessageParam): number {
-    return this.currentClient.estimateMessageTokens(message)
   }
 }
